Add ability to handle long scale
Given that all of the arrays except for BIG_PREFIXES use the English names for numbers, I realize that this library is intended to be used for English language applications.  MOST English speaking locales use short scale where billion means 10^9.  However, there are some locales that use long scale where billion means 10^12 and 10^9 is represented by the word milliard (and 10^15 is billion instead of trillion).

https://en.wikipedia.org/wiki/Long_and_short_scales